refactor(back): 抽象不同 jwt 的签发和验证以及中间件的实现, 并将 payload 挪到 ctx.state, to #8 #10
