+++ server/src/route/user/index.ts
@@ -3,11 +3,11 @@ import bodyParser from 'koa-bodyparser'
 import { register } from './register'
 import { update } from './update'
 import { query, queryMultiUser } from './query'
-import { jwtAuth } from '../../middleware/jwtAuth'
+import { UserLoginedAuthMiddleware } from '../../middleware/jwtAuth'
 import { Context, DefaultState } from 'koa'
 
 export const userRouter = new Router<DefaultState, Context>()
     .post('/user', bodyParser(), register)
-    .put('/user', jwtAuth, bodyParser(), update)
+    .put('/user', UserLoginedAuthMiddleware, bodyParser(), update)
     .get('/user', query)
     .get('/users', queryMultiUser)
